% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliometRics.R
\name{bibliometric}
\alias{bibliometric}
\title{Calculate Bibliometrics}
\usage{
bibliometric(bib, base = read.baselines(), discipline = NULL)
}
\arguments{
\item{bib}{A list with an element called  \code{pubs}, which is a
\code{data.frame} with columns \code{'Publication Year'} and
\code{'Total Citations'}.  Usually
from \code{\link{read.wos}}}

\item{base}{Baseline percentages}

\item{discipline}{Name of Discipline, must be in column names of
\code{base}}
}
\value{
A \code{data.frame} of metric
}
\description{
Calculate Bibliometrics
}
\examples{
wos_file = system.file("sbegueria.txt", package = "bibliometRics")
bib <- read.wos(wos_file)
bib$author = gsub('[ñÑ]', 'N', bib$author)
bib$pubs$Authors = gsub('[ñÑ]', 'N', bib$pubs$Authors)

base = read.baselines()
discipline = 'GEOSCIENCES'
bibliometric(bib, base, 'GEOSCIENCES')
}
