% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{biblioplot}
\alias{biblioplot}
\alias{biblioplot2}
\title{Plot Bibliometrics}
\usage{
biblioplot(bib, quant)

biblioplot2(bib)
}
\arguments{
\item{bib}{A list with an element called  \code{pubs}, which is a
\code{data.frame} with columns \code{'Publication Year'} and
\code{'Total Citations'}.  Usually
from \code{\link{read.wos}}}

\item{quant}{Quantile, passed to \code{\link{citation_rank}}}
}
\description{
Plot Bibliometrics
}
\examples{
wos_file = system.file("sbegueria.txt", package = "bibliometRics")
bib <- read.wos(wos_file)
base = read.baselines()
biblioplot(bib, q=base$GEOSCIENCES)
bib$author = gsub('Ñ', 'N', bib$author)
}
