% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_rank.R
\name{citation_rank}
\alias{citation_rank}
\title{Calculate citation ranking (quantiles)}
\usage{
citation_rank(x, w = NULL, quant = NULL)
}
\arguments{
\item{x}{A list with an element called  \code{pubs}, which is a
\code{data.frame} with columns \code{'Publication Year'} and
\code{'Total Citations'}.  Usually
from \code{\link{read.wos}}}

\item{w}{column indicators of which quantiles to grab}

\item{quant}{A \code{data.frame}? maybe}
}
\value{
A scalar number
}
\description{
Calculate citation ranking (quantiles)
}
\examples{
wos_file = system.file("sbegueria.txt", package = "bibliometRics")
bib <- read.wos(wos_file)
hirsch(bib)
hirsch(bib, 2010)
egghe(bib)
egghe(bib, 2010)
base = read.baselines()
citation_rank(bib, quant=base$GEOSCIENCES)
}
