% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{credentials}
\alias{credentials}
\title{Método para guardar credenciales para usar en Scopus}
\usage{
credentials()
}
\value{
Este método no devuelve nada
}
\description{
Método para guardar credenciales para usar en Scopus
}
\examples{
credentials()
}
