% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliometrics.R
\name{.onLoad}
\alias{.onLoad}
\title{Instalación por defecto de TinyTeX}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{pkgname}{}
}
\value{
nada
}
\description{
Instalación por defecto de TinyTeX
}
