% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArticles.R
\name{getArticle}
\alias{getArticle}
\title{Método para obtener los artículos en cualquier base de datos posible, mediante una consulta}
\usage{
getArticle(apis, query)
}
\arguments{
\item{apis}{Vector de las diferentes bases de datos, asignándole TRUE o FALSE, según la base de datos que queramos buscar.}

\item{query}{Elemento de texto, por el cual consultaremos en las bases de datos que son elegidas en apis.}
}
\value{
Este método, devuelve un dataframe de los artículos de que se obtienen de las diferentes bases de datos bibliométricas.
}
\description{
Método para obtener los artículos en cualquier base de datos posible, mediante una consulta
}
\examples{
getArticle(c(wos=TRUE,scopus=TRUE, scholar=FALSE),"tutorial autoencoder")
getArticle(c(wos=FALSE,scopus=TRUE, scholar=TRUE),"Strategies for time series forecasting with generalized regression neural networks")
}
