% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArticles.R
\name{getArticlesGoogle}
\alias{getArticlesGoogle}
\title{Método para obtener los artículos de la base de datos de Google Scholar}
\usage{
getArticlesGoogle(query, apiSelect)
}
\arguments{
\item{query}{Elemento de texto, por el cual consultaremos en la base de datos de Google Scholar.}

\item{apiSelect}{Pasa una variable JSON, tal que:
{
"name":"scholar",
"urlArticle": "https://serpapi.com/search.json?engine=google_scholar",
"key": clave de la API de google (serpapi)
}}
}
\value{
Este método, tras los parámetros facilitados, genera un objeto data.frame con los datos en los que coincide con la consulta obtenida por Google Scholar, que posteriormente será devuelto.
}
\description{
Método para obtener los artículos de la base de datos de Google Scholar
}
\examples{

getArticlesGoogle("tutorial encoder", "{"name":"scholar",
     "urlArticle": "https://serpapi.com/search.json?engine=google_scholar",
     "key": clave de la API de google (serpapi)}")

getArticlesGoogle("Strategies for time series forecasting with generalized regression neural networks", "{"name":"scholar",
     "urlArticle": "https://serpapi.com/search.json?engine=google_scholar",
     "key": clave de la API de google (serpapi)}")

}
