% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArticles.R
\name{getArticlesScopus}
\alias{getArticlesScopus}
\title{Método para obtener los artículos de la base de datos de Scopus}
\usage{
getArticlesScopus(query, apiSelect)
}
\arguments{
\item{query}{Elemento de texto, por el cual consultaremos en la base de datos de Scopus.}

\item{apiSelect}{Pasa una variable JSON, tal que:
{
"name":"scopus",
"urlArticle": "https://api.elsevier.com/content/search/scopus",
"key": Clave de la API de Scopus,
"instant": Esta variable "no sería necesaria", pero la uso, ya que me permitieron acceder con esta variable a APIs que no podía acceder con un usuario normal.
}}
}
\value{
Este método, tras los parámetros facilitados, genera un objeto data.frame con los datos en los que coincide con la consulta obtenida por Scopus, que posteriormente será devuelto.
}
\description{
Método para obtener los artículos de la base de datos de Scopus
}
\examples{

getArticlesScopus("tutorial encoder", {
"name":"scopus",
"urlArticle": "https://api.elsevier.com/content/search/scopus",
"key": Clave de la API de Scopus,
"instant": Esta variable "no sería necesaria", pero la uso, ya que me permitieron acceder con esta variable a APIs que no podía acceder con un usuario normal.
})

getArticlesScopus("Strategies for time series forecasting with generalized regression neural networks", {
"name":"scopus",
"urlArticle": "https://api.elsevier.com/content/search/scopus",
"key": Clave de la API de Scopus,
"instant": Esta variable "no sería necesaria", pero la uso, ya que me permitieron acceder con esta variable a APIs que no podía acceder con un usuario normal.
})
}
