% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAuthors.R
\name{getAuthor}
\alias{getAuthor}
\title{función para obtener articulos}
\usage{
getAuthor(apis, query)
}
\arguments{
\item{apis}{lista de apis diciendo con los valores de apis = true o false}

\item{query}{texto que queremos buscar
TENGO QUE ESCRIBIR ESTO}
}
\value{
devuelve dataframe de los artículos de las apis a buscar
}
\description{
función para obtener articulos
}
\examples{
getArticle(c(wos=TRUE,scopus=TRUE, scholar=FALSE),"springer")
getArticle(c(wos=FALSE,scopus=TRUE, scholar=TRUE),"charte")
}
