% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAuthors.R
\name{getAuthors}
\alias{getAuthors}
\title{Método para obtener los autores que coincidan con una consulta de texto}
\usage{
getAuthors(apis, query)
}
\arguments{
\item{apis}{Vector de las diferentes bases de datos, asignándole TRUE o FALSE, según la base de datos que queramos buscar.}

\item{query}{Elemento de texto, por el cual consultaremos en las bases de datos que son elegidas en apis.}
}
\value{
Este método, devuelve un dataframe de los autores que se obtienen de las diferentes bases de datos bibliométricas.
}
\description{
Método para obtener los autores que coincidan con una consulta de texto
}
\examples{
getAuthors(c(wos=TRUE,scopus=TRUE, scholar=FALSE),"Charte")
getAuthors(c(wos=FALSE,scopus=TRUE, scholar=TRUE),"Jiménez")
}
