% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAuthors.R
\name{getAuthorScopus}
\alias{getAuthorScopus}
\title{palabras clave' = NA,
autor/es' = result$'search-results'$entry$\code{dc:creator},
año' = result$'search-results'$entry$\code{prism:coverDate},
nombre fuente' = NA,
ISSN' = result$'search-results'$entry$\code{prism:issn},}
\usage{
getAuthorScopus(query, apiSelect)
}
\description{
palabras clave' = NA,
autor/es' = result$'search-results'$entry$\code{dc:creator},
año' = result$'search-results'$entry$\code{prism:coverDate},
nombre fuente' = NA,
ISSN' = result$'search-results'$entry$\code{prism:issn},
}
