% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAuthors.R
\name{getAuthorsGoogle}
\alias{getAuthorsGoogle}
\title{Método para obtener los autores de la base de datos de Google Scholar}
\usage{
getAuthorsGoogle(query)
}
\arguments{
\item{query}{Elemento de texto, por el cual consultaremos en las bases de datos que son elegidas en apis.}
}
\value{
Este método, tras los parámetros facilitados, genera un objeto data.frame con los datos en los que coincide con la consulta obtenida por Google Scholar, que posteriormente será devuelto.
}
\description{
Método para obtener los autores de la base de datos de Google Scholar
}
\examples{
getAuthorsGoogle("Charte")
getAuthorsGoogle("Jiménez")
}
