% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetricsArticle.R
\name{getMetricsArticle}
\alias{getMetricsArticle}
\title{Método para obtener métricas de un artículo de las diferentes bases de datos bibliométricas.}
\usage{
getMetricsArticle(apis, query)
}
\arguments{
\item{apis}{Nombre de la base de datos a buscar}

\item{query}{En este caso será la consulta será el ID del artículo.}
}
\value{
Este método, devuelve un dataframe de las métricas que se pueden obtener de un artículo de las diferentes bases de datos bibliométricas.
}
\description{
Método para obtener métricas de un artículo de las diferentes bases de datos bibliométricas.
}
\examples{
getMetricsSource("scopus","1874-9305", NA)
getMetricsSource("scopus",NA, "Elsevier Astrodynamics Series")
}
