% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetricsAuthor.R
\name{getMetricsAuthor}
\alias{getMetricsAuthor}
\title{Método para obtener métricas de un autor de las diferentes bases de datos bibliométricas.}
\usage{
getMetricsAuthor(apis, ID)
}
\arguments{
\item{apis}{La variable apis será el nombre en texto que tenemos asociado a cada base de datos bibliométrica}

\item{ID}{Variable de texto que tendremos asociado el ID del autor, para buscarlo.}
}
\value{
Este método devuelve un dataframe en el cual estará los datos métricos del autor solicitado por el ID de la base de datos escogida.
}
\description{
Método para obtener métricas de un autor de las diferentes bases de datos bibliométricas.
}
\examples{
getMetricsAuthor("scholar","i8l_80EAAAAJ")
getMetricsAuthor("scopus","9-s2.0-40661023100")
}
