% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetricsAuthorScholar.R
\name{getMetricsAuthorScholar}
\alias{getMetricsAuthorScholar}
\title{Obtener métricas de Autores de Google Scholar}
\usage{
getMetricsAuthorScholar(uid)
}
\arguments{
\item{uid}{id de Google Scholar}
}
\value{
dataframe de datos métricos
}
\description{
Obtener métricas de Autores de Google Scholar
}
\examples{
getMetricsAuthorScholar("i8l_80EAAAAJ")
}
