% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetricsAuthor.R
\name{getMetricsAuthorScopus}
\alias{getMetricsAuthorScopus}
\title{Método para obtener métricas de un autor de la base de datos bibliométrica de Scopus.}
\usage{
getMetricsAuthorScopus(ID, apis)
}
\arguments{
\item{ID}{Número de serie por el cual consultaremos en la base de datos de Scopus.}

\item{apis}{Nombre asignado a la base de datos, siempre será "scopus" si entra en este método.}
}
\value{
Este método devuelve un dataframe en el cual estará los datos métricos del autor solicitado por el ID de la base de datos Scopus.
}
\description{
Método para obtener métricas de un autor de la base de datos bibliométrica de Scopus.
}
\examples{
getMetricsAuthorScopus("9-s2.0-40661023100","scopus")

}
