% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetrics.R
\name{getMetricsScopus}
\alias{getMetricsScopus}
\title{Obtener Métricas de Scopus mediante web scrapping}
\usage{
getMetricsScopus(uid)
}
\arguments{
\item{uid}{del artículo}
}
\value{
diccionario con métricas
}
\description{
Obtener Métricas de Scopus mediante web scrapping
}
\examples{
getMetricsScopus("2-s2.0-85187374370")
}
