% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetricsSource.R
\name{getMetricsSource}
\alias{getMetricsSource}
\title{Método para obtener métricas de una revista de las diferentes bases de datos bibliométricas.}
\usage{
getMetricsSource(apis, query, title)
}
\arguments{
\item{apis}{Nombre de la base de datos a buscar}

\item{query}{En este caso será la consulta será el ISSN de la revista.}

\item{title}{En caso de que no se inserte la ISSN, se podrá buscar por el nombre de la revista.}
}
\value{
Este método, devuelve una lista de dataframes de las métricas que se pueden obtener de una revista de las diferentes bases de datos bibliométricas.
}
\description{
Método para obtener métricas de una revista de las diferentes bases de datos bibliométricas.
}
\examples{
getMetricsSource("scopus","1874-9305", NA)
getMetricsSource("scopus",NA, "Elsevier Astrodynamics Series")
}
