% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSource.R
\name{getSources}
\alias{getSources}
\title{Método para obtener la información de todas las revistas}
\usage{
getSources(apis, query)
}
\arguments{
\item{apis}{Vector de las diferentes bases de datos, asignándole TRUE o FALSE, según si la base de datos que queramos buscar.}

\item{query}{Elemento de texto, por el cual consultaremos en las bases de datos que son elegidas en apis.}
}
\value{
Este método, devuelve un dataframe de las revistas que se obtienen de las diferentes bases de datos bibliométricas.
}
\description{
Método para obtener la información de todas las revistas
}
\examples{
getSources(c(wos=FALSE,scopus=TRUE, scholar=FALSE),"elsevier")
getSources(c(wos=FALSE,scopus=TRUE, scholar=FALSE),"neurocomputing")
}
