% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSource.R
\name{getSourceScopus}
\alias{getSourceScopus}
\title{Método para obtener la información de revistas de la base de datos de Scopus}
\usage{
getSourceScopus(query, apiSelect)
}
\arguments{
\item{query}{Elemento de texto, por el cual consultaremos en las bases de datos que son elegidas en apis.}

\item{apis}{Vector de las diferentes bases de datos, asignándole TRUE o FALSE, según la base de datos que queramos buscar.}
}
\value{
Este método, tras los parámetros facilitados, genera un objeto data.frame con los datos en los que coincide con la consulta obtenida por Scopus, que posteriormente será devuelto.
}
\description{
Método para obtener la información de revistas de la base de datos de Scopus
}
\examples{
getSourceScopus("scopus","elsevier")
getSourceScopus("scopus","neurocomputing")
}
