% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.baselines.R
\name{read.baselines}
\alias{read.baselines}
\title{Read field percentile baselines (FPB) from the Essential Science
Indicators database.}
\usage{
read.baselines(infile = system.file("BaselinePercentiles.csv", package =
  "bibliometRics"))
}
\arguments{
\item{infile}{Baseline percentages for each field}
}
\value{
A \code{list} of percentages
}
\description{
Read field percentile baselines (FPB) from the Essential Science
Indicators database.
}
\examples{
res = read.baselines()
}
