% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.wos.R
\name{read.wos}
\alias{read.wos}
\title{Read and process Web of Science citation report data.}
\usage{
read.wos(infile)
}
\arguments{
\item{infile}{Input text file from Web of Science}
}
\value{
A list of information
}
\description{
Read and process Web of Science citation report data.
}
\examples{
wos_file = system.file('sbegueria.txt', package = "bibliometRics")
bib <- read.wos(wos_file)
}
