% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnviromentVariables.R
\name{setEmail}
\alias{setEmail}
\title{almacena email corporativo para poder acceder a recursos}
\usage{
setEmail(value)
}
\arguments{
\item{value}{debe ir el correo electrónico corporativo
TENGO QUE ESCRIBIR ESTO}
}
\value{
None
}
\description{
almacena email corporativo para poder acceder a recursos
}
\examples{
setEmail("prueba@red.ujaen.es")
}
