% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnviromentVariables.R
\name{setPass}
\alias{setPass}
\title{almacena contraseña de la cuenta corporativa}
\usage{
setPass(value)
}
\arguments{
\item{value}{debe ir contraseña a guardar}
}
\value{
None
}
\description{
almacena contraseña de la cuenta corporativa
}
\examples{
setPass("prueba")
setPass("contrasena")
}
