% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnviromentVariables.R
\name{setPython}
\alias{setPython}
\title{almacena ruta de python}
\usage{
setPython(value)
}
\arguments{
\item{value}{debe ser la ruta del directorio que aloje el ejecutable python}
}
\value{
None
}
\description{
almacena ruta de python
}
\examples{
setPython("C:/User/Python")
}
