% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{setup_renviron}
\alias{setup_renviron}
\title{Método para guardar credenciales para usar en Scopus}
\usage{
setup_renviron()
}
\value{
Este método no devuelve nada
}
\description{
Método para guardar credenciales para usar en Scopus
}
\examples{
setup_renviron()
}
