% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnviromentVariables.R
\name{setUser}
\alias{setUser}
\title{almacena email corporativo para poder acceder a recursos}
\usage{
setUser(value)
}
\arguments{
\item{value}{debe ir el correo electrónico corporativo
TENGO QUE ESCRIBIR ESTO}
}
\value{
None
}
\description{
almacena email corporativo para poder acceder a recursos
}
\examples{
setEmail("prueba")
}
