% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prueba.R
\name{suma}
\alias{suma}
\title{Método SUMA}
\usage{
suma(a, b)
}
\arguments{
\item{a}{primer número}

\item{b}{segundo número}
}
\value{
resultado de la suma
}
\description{
Método SUMA
}
\examples{
suma(a=6, b=6)
suma(b=7, a=3)
}
