% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetricsAuthor.R
\name{getMetricsAuthorScholar}
\alias{getMetricsAuthorScholar}
\title{Método para obtener métricas de un autor de la base de datos bibliométrica Google Scholar.}
\usage{
getMetricsAuthorScholar(uid)
}
\arguments{
\item{uid}{id de Google Scholar}
}
\value{
Este método devuelve un dataframe en el cual estará los datos métricos del autor solicitado por el ID de la base de datos Google Scholar.
}
\description{
Método para obtener métricas de un autor de la base de datos bibliométrica Google Scholar.
}
\examples{
getMetricsAuthorScholar("i8l_80EAAAAJ")
getMetricsAuthorScholar("iwXvlSAAAAAJ")
}
