% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetricsAuthorScopus.R
\name{getMetricsAuthorScopus}
\alias{getMetricsAuthorScopus}
\title{Obtener métricas de Autores de Scopus}
\usage{
getMetricsAuthorScopus(query, apis)
}
\arguments{
\item{uid}{id de Scopus}
}
\value{
dataframe de datos métricos
}
\description{
Obtener métricas de Autores de Scopus
}
\examples{
getMetricsAuthorScopus("9-s2.0-40661023100","scopus")
}
