% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetricsArticle.R
\name{getMetricsScopus}
\alias{getMetricsScopus}
\title{Método para obtener métricas de Scopus de un artículo mediante web scrapping}
\usage{
getMetricsScopus(uid)
}
\arguments{
\item{uid}{Número de serie UID del artículo}
}
\value{
Este método, tras pasar el UID del artículo y haber introducido en el archivo .Renviron, genera un objeto data.frame con las métricas del artículo con dicho UID en Scopus, que posteriormente será devuelto.
}
\description{
Método para obtener métricas de Scopus de un artículo mediante web scrapping
}
\examples{
getMetricsScopus("2-s2.0-85187374370")
}
