% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetricsSource.R
\name{getMetricsSource}
\alias{getMetricsSource}
\title{Obtener métricas de fuentes de Scopus}
\usage{
getMetricsSource(apis, query)
}
\arguments{
\item{apis}{BBDD scopus}

\item{query}{consulta}
}
\value{
dataframe de datos métricos
}
\description{
Obtener métricas de fuentes de Scopus
}
\examples{
getMetricsSource("scopus","9-s2.0-40661023100")
}
