% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSource.R
\name{getSource}
\alias{getSource}
\title{función para obtener información de fuentes}
\usage{
getSource(apis, query)
}
\arguments{
\item{apis}{lista de apis diciendo con los valores de apis = true o false}

\item{query}{texto que queremos buscar
TENGO QUE ESCRIBIR ESTO}
}
\value{
devuelve dataframe de la información de las fuentes de las apis a buscar
}
\description{
función para obtener información de fuentes
}
\examples{
getSource(c(wos=TRUE,scopus=TRUE, scholar=FALSE),"springer")
getSource(c(wos=FALSE,scopus=TRUE, scholar=TRUE),"charte")
}
